Replace foo_role_packages in examples by foo_packages
It is not clear if in a role named `bar` the variable to hold packages should be
called `bar_packages` (`foo_role` replaced by the role name `bar`) or
`bar_role_packages` (`foo` replaced by `bar`). Make this consistent by
omitting `role_` from the example variable names, this making clear that
the literal string `role_` is not supposed to be a part of the variable name.